# api.yaml
openapi: 3.0.0
info:
  title: Astro Clock API
  version: 1.0.0
  description: API for astronomical chart calculations and job management
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /chart:
    get:
      summary: Generate astrological chart
      description: |
        Generates and returns an astrological chart image. Supported output
        formats are png and svg. For structured JSON chart data, use
        /api/v1/chart/data.
      operationId: getChart
      parameters:
        - name: lat
          in: query
          description: Latitude of the location
          schema:
            type: number
            format: double
            default: 0.0
        - name: lon
          in: query
          description: Longitude of the location
          schema:
            type: number
            format: double
            default: 0.0
        - name: time
          in: query
          description: Time for the chart (optional)
          schema:
            type: string
        - name: format
          in: query
          description: "Image output format. Supported values: png, svg. Defaults to png."
          schema:
            type: string
            enum: [png, svg]
            default: png
      responses:
        '200':
          description: Chart image output in the requested format
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/svg+xml:
              schema:
                type: string
        '500':
          description: Internal server error
  /api/v1/load:
    post:
      summary: Create load job
      description: |
        Creates and executes a data load job. 
        When sync=true, blocks until completion and returns result.
        When sync=false or omitted, returns immediately with job-id for polling.
      operationId: createLoadJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - start_date
                - days
              properties:
                start_date:
                  type: string
                  description: Start date in YYYY-MM-DD format
                  example: "2024-01-01"
                days:
                  type: integer
                  description: Number of days to load (1-365)
                  minimum: 1
                  maximum: 365
                  example: 7
                sync:
                  type: boolean
                  description: If true, execute synchronously and wait for completion
                  default: false
      responses:
        '200':
          description: Sync load completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadSyncResponse'
        '202':
          description: Async load job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadAsyncResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
  /api/v1/jobs/{id}:
    get:
      summary: Get job status
      description: Retrieves the details and status of a specific job
      operationId: getJob
      parameters:
        - name: id
          in: path
          required: true
          description: Job UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
    delete:
      summary: Delete a job
      description: |
        Deletes a single job by ID. Jobs with status `in_process` cannot be deleted
        (returns 409 Conflict) to prevent orphaning an active worker. Deleting a
        non-existent job returns 404. Deleting a job that disappears between the
        existence check and deletion returns 404 (idempotent).
      operationId: deleteJob
      responses:
        '204':
          description: Job deleted successfully (no content)
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Cannot delete job in in_process status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
  /api/v1/jobs:
    get:
      summary: List jobs
      description: |
        Returns a cursor-paginated list of jobs with multi-value filtering.
        Results are ordered by created_at DESC, id DESC. Cursor-based pagination
        uses opaque base64 tokens anchored on (created_at, id) for stable page
        boundaries — new jobs inserted between page fetches don't shift results.
      operationId: listJobs
      parameters:
        - name: status
          in: query
          description: |
            Filter by job status. Accepts comma-separated list for multi-value filtering.
            Invalid values return 400 with descriptive error.
          schema:
            type: string
          examples:
            single:
              value: complete
              summary: Single status filter
            multi:
              value: "complete,failed"
              summary: Multiple statuses
        - name: job_type
          in: query
          description: |
            Filter by job type. Accepts comma-separated list for multi-value filtering.
            Invalid values return 400 with descriptive error.
          schema:
            type: string
          examples:
            single:
              value: load
              summary: Single job type filter
            multi:
              value: "load,query"
              summary: Multiple job types
        - name: created_after
          in: query
          description: |
            Filter to jobs created on or after this timestamp.
            Accepts RFC3339 (e.g. 2025-01-15T00:00:00Z) or YYYY-MM-DD (treated as midnight UTC).
          schema:
            type: string
          example: "2025-01-01"
        - name: created_before
          in: query
          description: |
            Filter to jobs created on or before this timestamp.
            Accepts RFC3339 (e.g. 2025-03-01T23:59:59Z) or YYYY-MM-DD (treated as midnight UTC).
          schema:
            type: string
          example: "2025-03-01"
        - name: count
          in: query
          description: "Number of jobs to return per page (default 20, max 100)"
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          description: |
            Opaque cursor token for pagination. Obtain from the `next` or `prev` field
            of a previous response. Clients should treat this as an opaque string.
            Omit for the first page.
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListJobsResponse'
        '400':
          description: Invalid query parameters (invalid status/job_type values, invalid date format, or invalid cursor)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
  /api/v1/query/wedding:
    post:
      summary: Execute wedding query
      description: |
        Execute a wedding date query.
        When sync=true, blocks until completion and returns result.
        When sync=false or omitted, returns immediately with job-id for polling.
      operationId: queryWedding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Sync query completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuerySyncResponse'
        '202':
          description: Async query job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryAsyncResponse'
        '400':
          description: Invalid request parameters
        '500':
          description: Internal server error
  /api/v1/query/project:
    post:
      summary: Execute project query
      description: |
        Execute a project start date query.
        When sync=true, blocks until completion and returns result.
        When sync=false or omitted, returns immediately with job-id for polling.
      operationId: queryProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Sync query completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuerySyncResponse'
        '202':
          description: Async query job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryAsyncResponse'
        '400':
          description: Invalid request parameters
        '500':
          description: Internal server error
  /api/v1/query/travel:
    post:
      summary: Execute travel query
      description: |
        Execute a travel date query.
        When sync=true, blocks until completion and returns result.
        When sync=false or omitted, returns immediately with job-id for polling.
      operationId: queryTravel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Sync query completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuerySyncResponse'
        '202':
          description: Async query job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryAsyncResponse'
        '400':
          description: Invalid request parameters
        '500':
          description: Internal server error
  /api/v1/query/{query_name}:
    post:
      summary: Execute generic named query
      description: |
        Execute a named query (wedding, project, or travel).
        When sync=true, blocks until completion and returns result.
        When sync=false or omitted, returns immediately with job-id for polling.
      operationId: executeQuery
      parameters:
        - name: query_name
          in: path
          required: true
          description: Name of the query to execute
          schema:
            type: string
            enum: [wedding, project, travel]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Sync query completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuerySyncResponse'
        '202':
          description: Async query job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryAsyncResponse'
        '400':
          description: Invalid request parameters
        '500':
          description: Internal server error
  /api/v1/chart/data:
    get:
      summary: Get chart data as JSON
      description: |
        Returns structured astrological chart data including planet positions,
        house cusps, aspects, grand trines, and moon void of course information.
        Suitable for display in external applications.
      operationId: getChartData
      parameters:
        - name: lat
          in: query
          required: true
          description: Latitude of the location (-90 to 90)
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: lon
          in: query
          required: true
          description: Longitude of the location (-180 to 180)
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: time
          in: query
          required: false
          description: Time for the chart (optional, defaults to current time)
          schema:
            type: string
        - name: house
          in: query
          required: false
          description: House system for calculating house cusps. Defaults to Placidus.
          schema:
            type: string
            enum:
              - Placidus
              - Koch
              - Equal
              - Whole
              - Porphyry
              - Regiomontanus
              - Campanus
              - Morinus
              - Alcabitus
              - Topocentric
              - Vehlow
            default: Placidus
      responses:
        '200':
          description: Chart data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChartDataResponse'
        '400':
          description: Invalid request parameters (missing or invalid lat/lon)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error during chart calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    LoadSyncResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          example: complete
        result:
          type: object
          description: Job result data (present when complete)
    LoadAsyncResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          example: pending
        poll_url:
          type: string
          description: URL to poll for job status
          example: "/api/v1/jobs/00000000-0000-0000-0000-000000000000"
    QueryRequest:
      type: object
      required:
        - start_date
      properties:
        start_date:
          type: string
          description: Start date in YYYY-MM-DD format
          example: "2024-01-01"
        days:
          type: integer
          description: Number of days to query (1-366). Mutually exclusive with end_date.
          minimum: 1
          maximum: 366
          example: 30
        end_date:
          type: string
          description: End date in YYYY-MM-DD format. Mutually exclusive with days.
          example: "2024-01-30"
        sync:
          type: boolean
          description: Execute synchronously
          default: false
...
    JobResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        job_type:
          type: string
          example: load
        status:
          type: string
          enum: [pending, in_process, complete, failed]
        payload:
          type: object
          description: Job input parameters
        result:
          type: object
          description: Job output (present when complete)
        error:
          type: object
          description: Error details (present when failed)
          properties:
            code:
              type: string
            message:
              type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code
          example: invalid_date
        message:
          type: string
          description: Human-readable error message